requirements.txt
anthropic
pandas
openpyxl
python-dotenv
jinja2
groq